requirement.txt
absl-py==0.8.0
h5py==2.8.0
matplotlib==3.0.3
numpy==1.14.5
opencv-python==4.0.0.21
Pillow==6.0.0
protobuf==3.7.1
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.4.1
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0

